+++ src/app/view-config-versions.js
@@ -14,7 +14,7 @@ import configSchema1_0_9 from '../schemas/config-1.0.9.schema.json';
 import configSchema1_0_10 from '../schemas/config-1.0.10.schema.json';
 import configSchema1_0_11 from '../schemas/config-1.0.11.schema.json';
 import configSchema1_0_12 from '../schemas/config-1.0.12.schema.json';
-import cellSetsSchema from '../schemas/cell-sets.schema.json';
+import obsSetsSchema from '../schemas/obsSets.schema.json';
 import rasterSchema from '../schemas/raster.schema.json';
 import {
   upgradeFrom0_1_0,
@@ -40,17 +40,17 @@ import {
 export const LATEST_VERSION = '1.0.12';
 export const SCHEMA_HANDLERS = {
   '0.1.0': [new Ajv().compile(configSchema0_1_0), upgradeFrom0_1_0],
-  '1.0.0': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_0), upgradeFrom1_0_0],
-  '1.0.1': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_1), upgradeFrom1_0_1],
-  '1.0.2': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_2), upgradeFrom1_0_2],
-  '1.0.3': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_3), upgradeFrom1_0_3],
-  '1.0.4': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_4), upgradeFrom1_0_4],
-  '1.0.5': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_5), upgradeFrom1_0_5],
-  '1.0.6': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_6), upgradeFrom1_0_6],
-  '1.0.7': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_7), upgradeFrom1_0_7],
-  '1.0.8': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_8), upgradeFrom1_0_8],
-  '1.0.9': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_9), upgradeFrom1_0_9],
-  '1.0.10': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_10), upgradeFrom1_0_10],
-  '1.0.11': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_11), upgradeFrom1_0_11],
-  '1.0.12': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_12), null],
+  '1.0.0': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_0), upgradeFrom1_0_0],
+  '1.0.1': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_1), upgradeFrom1_0_1],
+  '1.0.2': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_2), upgradeFrom1_0_2],
+  '1.0.3': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_3), upgradeFrom1_0_3],
+  '1.0.4': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_4), upgradeFrom1_0_4],
+  '1.0.5': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_5), upgradeFrom1_0_5],
+  '1.0.6': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_6), upgradeFrom1_0_6],
+  '1.0.7': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_7), upgradeFrom1_0_7],
+  '1.0.8': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_8), upgradeFrom1_0_8],
+  '1.0.9': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_9), upgradeFrom1_0_9],
+  '1.0.10': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_10), upgradeFrom1_0_10],
+  '1.0.11': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_11), upgradeFrom1_0_11],
+  '1.0.12': [new Ajv().addSchema(obsSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_12), null],
 };