+++ src/loaders/AbstractLoader.js
@@ -6,8 +6,10 @@ import uuidv4 from 'uuid/v4';
  */
 export default class AbstractLoader {
   constructor({
-    type, url, requestInit, options,
+    fileType, type,
+    url, requestInit, options,
   }) {
+    this.fileType = fileType;
     this.type = type;
     this.url = url;
     this.requestInit = requestInit;