+++ src/loaders/ObsSetsJsonLoader.js
@@ -1,4 +1,4 @@
-import cellSetsSchema from '../schemas/cell-sets.schema.json';
+import obsSetsSchema from '../schemas/obsSets.schema.json';
 import JsonLoader from './JsonLoader';
 import { tryUpgradeTreeToLatestSchema } from '../components/sets/io';
 import { AbstractLoaderError } from './errors';
@@ -6,11 +6,11 @@ import LoaderResult from './LoaderResult';
 
 import { initializeCellSetColor } from '../components/sets/cell-set-utils';
 
-export default class CellSetsJsonLoader extends JsonLoader {
+export default class ObsSetsJsonLoader extends JsonLoader {
   constructor(dataSource, params) {
     super(dataSource, params);
 
-    this.schema = cellSetsSchema;
+    this.schema = obsSetsSchema;
   }
 
   async load() {
@@ -19,7 +19,8 @@ export default class CellSetsJsonLoader extends JsonLoader {
       return Promise.reject(payload);
     }
     const { data: rawData, url } = payload;
-    const upgradedData = tryUpgradeTreeToLatestSchema(rawData, 'cell');
+    const datatype = this.fileType.endsWith('cell-sets.json') ? 'cell' : 'obs';
+    const upgradedData = tryUpgradeTreeToLatestSchema(rawData, datatype);
 
     const coordinationValues = {
       obsSetSelection: [],