+++ src/loaders/types.js
@@ -9,12 +9,14 @@ import GenesJsonAsMatrixZarrLoader from './GenesJsonAsMatrixZarrLoader';
 import ClustersJsonAsMatrixZarrLoader from './ClustersJsonAsMatrixZarrLoader';
 import RasterJsonLoader from './RasterJsonLoader';
 import OmeZarrLoader from './OmeZarrLoader';
-import CellSetsJsonLoader from './CellSetsJsonLoader';
+import ObsSetsJsonLoader from './ObsSetsJsonLoader';
 import AnnDataLoaders from './anndata-loaders';
 import GenomicProfilesZarrLoader from './GenomicProfilesZarrLoader';
 import { AnnDataSource, ZarrDataSource, JsonSource } from './data-sources';
 
 export const fileTypeToLoaderAndSource = {
+  [FileType.OBS_SETS_JSON]: [JsonSource, ObsSetsJsonLoader],
+  [FileType.OBS_SETS_CELL_SETS_JSON]: [JsonSource, ObsSetsJsonLoader],
   [FileType.EXPRESSION_MATRIX_ZARR]: [ZarrDataSource, MatrixZarrLoader],
   [FileType.CLUSTERS_JSON]: [JsonSource, ClustersJsonAsMatrixZarrLoader],
   [FileType.GENES_JSON]: [JsonSource, GenesJsonAsMatrixZarrLoader],
@@ -23,7 +25,7 @@ export const fileTypeToLoaderAndSource = {
   [FileType.NEIGHBORHOODS_JSON]: [JsonSource, JsonLoader],
   [FileType.RASTER_JSON]: [JsonSource, RasterJsonLoader],
   [FileType.RASTER_OME_ZARR]: [ZarrDataSource, OmeZarrLoader],
-  [FileType.CELL_SETS_JSON]: [JsonSource, CellSetsJsonLoader],
+  [FileType.CELL_SETS_JSON]: [JsonSource, ObsSetsJsonLoader],
   [FileType.ANNDATA_CELL_SETS_ZARR]: [AnnDataSource, AnnDataLoaders.CellSetsZarrLoader],
   [FileType.ANNDATA_CELLS_ZARR]: [AnnDataSource, AnnDataLoaders.CellsZarrLoader],
   [FileType.ANNDATA_EXPRESSION_MATRIX_ZARR]: [AnnDataSource, AnnDataLoaders.MatrixZarrLoader],